Restructured database module
Uniform module setup improves readability. Moved clients into subdirectory, added init files, and renamed Client classes to be simpler and more direct.
