+++ loom/database/__init__.py
@@ -1,1 +1,2 @@
-from .mongodb_clients import LoomMongoDBMotorTornadoClient, LoomMongoDBMotorAsyncioClient
+from .clients import MongoDBMotorTornadoClient, MongoDBMotorAsyncioClient
+from .interfaces import MongoDBTornadoInterface, MongoDBAsyncioInterface
