+++ loom/database/interfaces/__init__.py
@@ -1,2 +1,2 @@
-from .abstract import AbstractDBInterface
-from .mongodb import MongoDBTornadoInterface, MongoDBAsyncioInterface
+from .abstract_interface import AbstractDBInterface
+from .mongodb_interfaces import MongoDBTornadoInterface, MongoDBAsyncioInterface
